requirements.txt
altgraph==0.17.5
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
idna==3.11
iniconfig==2.3.0
macholib==1.16.4
Markdown==3.10
openpyxl==3.1.5
packaging==25.0
pluggy==1.6.0
portalocker==3.2.0
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
pytest==9.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
qtwidgets==1.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
urllib3==2.6.2
waitress==3.0.2
